Complete test for 'instance_of' predicate factory
