+++ tests/typology_tests/instance_of_tests/test_properties.py
@@ -1,6 +1,4 @@
-import inspect
-from typing import (Any,
-                    Sequence)
+from typing import Any
 
 from lz.typology import instance_of
 from tests.utils import equivalence
@@ -10,14 +8,6 @@ def test_type(class_: type) -> None:
     result = instance_of(class_)
 
     assert callable(result)
-
-
-def test_signature(class_: type) -> None:
-    result = instance_of(class_)
-
-    signature = inspect.signature(result)
-
-    assert len(signature.parameters) == 1
 
 
 def test_result_type(class_: type,
@@ -40,18 +30,3 @@ def test_commutativity(class_: type, other_class: type, object_: Any) -> None:
     right_function = instance_of(other_class, class_)
 
     assert equivalence(left_function(object_), right_function(object_))
-
-
-def test_base_case(object_: Any) -> None:
-    result = instance_of()
-
-    assert not result(object_)
-
-
-def test_step(classes: Sequence[type], class_: type, object_: Any) -> None:
-    base_function = instance_of(*classes)
-    adjunct_function = instance_of(class_)
-    summary_function = instance_of(*classes, class_)
-
-    assert equivalence(summary_function(object_),
-                       base_function(object_) or adjunct_function(object_))
